Удали ненужный код зависящий от data файла
